S-14c extract recall and background-tasks pages to web/modules/recall-bgtasks.js
Move loadRecallPage/doRecall/renderRecall* (~168 lines) and
loadBgTasksPage/startBgTask/loadBgTasks/viewBgTask/cancelBgTask (~171 lines)
from app.js into recall-bgtasks.js. initRecallBgTasks() exposes viewBgTask and
cancelBgTask as window globals for the inline onclick handlers in the bg-tasks
list renderer. Also fix a shadowing bug in loadBgTasks: the original code used
't' as both the map-callback parameter (task object) and the i18n function,
causing status labels to fail when called; renamed to 'task'.
